Refactor - BasicInfo 섹션에서 Categories 컴포넌트 분리
