+++ src/Components/BasicInfo/BasicInfo.jsx
@@ -1,73 +1,20 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { withCurlyBraces, paddingClassNameGenerator } from 'utils';
-import { categories as originCategory, productCode, stock } from 'data/basicInfo';
-import { A11yInput, CheckBox, Table, TableHeader, TableRow } from 'Components';
-import Card from './Card/Card';
-import Grid from 'Components/Grid/Grid';
-import Button from './Button/Button';
+import { productCode, stock } from 'data/basicInfo';
+import { A11yInput, Table, TableHeader, TableRow } from 'Components';
 import SearchFilter from './SearchFilter/SearchFilter';
 import styles from './BasicInfo.module.css';
 import commonStyles from 'styles/index.module.scss';
+import Categories from './Categories/Categories';
 
 const BasicInfo = () => {
-  const [categories, setCheckedCategories] = useState([...originCategory]);
-
-  const toggleCheckState = category => {
-    const newState = [...categories];
-    let id = categories.findIndex(({ id }) => id === category.id);
-    newState[id].checked = !category.checked;
-    setCheckedCategories(newState);
-  };
-
-  const checkedCategories = () => {
-    return categories.filter(({ checked }) => checked);
-  };
-
   return (
     <Table pdLevel={5}>
       <TableHeader className={commonStyles[paddingClassNameGenerator(11)]}>
         상품 기본 정보
       </TableHeader>
       <TableRow label="카테고리 *" pdLevel={0}>
-        <Grid container space={5}>
-          <Grid size={7} item>
-            <Card>
-              <ul className={styles.categories}>
-                {categories.map(category => {
-                  return (
-                    <CheckBox
-                      key={category.id}
-                      category={category.category}
-                      checked={category.checked}
-                      onChange={() => toggleCheckState(category)}
-                      name={`basicInfo-${category.category}`}
-                    >
-                      {withCurlyBraces(category.category)}
-                    </CheckBox>
-                  );
-                })}
-              </ul>
-            </Card>
-          </Grid>
-          <Grid size={5} item>
-            <Card>
-              <ul className={styles.checkedCategories}>
-                {checkedCategories().map(category => {
-                  return (
-                    <li key={category.id}>
-                      <Button type={1} onClick={() => toggleCheckState(category)}>
-                        {withCurlyBraces(category.category)} X
-                      </Button>
-                    </li>
-                  );
-                })}
-              </ul>
-              {checkedCategories().length === 0 && (
-                <span className={styles.message}>카테고리를 지정해 주세요.</span>
-              )}
-            </Card>
-          </Grid>
-        </Grid>
+        <Categories />
       </TableRow>
       <TableRow label="필터 태그" className={styles.relative}>
         <SearchFilter />
